+++ sdk/servicebus/azure-servicebus/samples/async_samples/send_queue_async.py
@@ -9,8 +9,6 @@
 Example to show sending message(s) to a Service Bus Queue asynchronously.
 """
 
-# pylint: disable=C0111
-
 import os
 import asyncio
 from azure.servicebus import ServiceBusMessage